bunny_sdk/bunny_api_client/pullzone/item/set_force_s_s_l/set_force_s_s_l_post_request_body.py
from __future__ import annotations
from dataclasses import dataclass, field
from kiota_abstractions.serialization import AdditionalDataHolder, Parsable, ParseNode, SerializationWriter
from typing import Any, Callable, Dict, List, Optional, TYPE_CHECKING, Union

@dataclass
class SetForceSSLPostRequestBody(AdditionalDataHolder, Parsable):
    # Stores additional data not described in the OpenAPI description found when deserializing. Can be used for serialization as well.
    additional_data: Dict[str, Any] = field(default_factory=dict)

    # Set to true to force SSL on the given pull zone hostname
    force_s_s_l: Optional[bool] = None
    # The hostname that will be updated
    hostname: Optional[str] = None
    
    @staticmethod
    def create_from_discriminator_value(parse_node: ParseNode) -> SetForceSSLPostRequestBody:
        """
        Creates a new instance of the appropriate class based on discriminator value
        param parse_node: The parse node to use to read the discriminator value and create the object
        Returns: SetForceSSLPostRequestBody
        """
        if parse_node is None:
            raise TypeError("parse_node cannot be null.")
        return SetForceSSLPostRequestBody()
    
    def get_field_deserializers(self,) -> Dict[str, Callable[[ParseNode], None]]:
        """
        The deserialization information for the current model
        Returns: Dict[str, Callable[[ParseNode], None]]
        """
        fields: Dict[str, Callable[[Any], None]] = {
            "ForceSSL": lambda n : setattr(self, 'force_s_s_l', n.get_bool_value()),
            "Hostname": lambda n : setattr(self, 'hostname', n.get_str_value()),
        }
        return fields
    
    def serialize(self,writer: SerializationWriter) -> None:
        """
        Serializes information the current object
        param writer: Serialization writer to use to serialize this model
        Returns: None
        """
        if writer is None:
            raise TypeError("writer cannot be null.")
        writer.write_bool_value("ForceSSL", self.force_s_s_l)
        writer.write_str_value("Hostname", self.hostname)
        writer.write_additional_data_value(self.additional_data)
    

