pyproject.toml
[tool.poetry]
name = "bunny-sdk"
version = "0.0.13"
description = "Python SDK for bunny.net 🐰"
authors = ["John L. Armstrong IV"]
license = "AGPL-3.0-only"
readme = "README.md"
keywords = [
  "BunnyCDN",
  "bunny.net",
  "bunny",
  "cdn",
  "bunny-launcher",
  "sdk",
  "api",
]

[tool.poetry.urls]
Homepage = "https://bunny-launcher.net/bunny-sdk/languages/python/"
Documentation = "https://bunny-launcher.net/bunny-sdk/languages/python/"
Repository = "https://github.com/jlarmstrongiv/bunny-sdk-python.git"
Issues = "https://github.com/jlarmstrongiv/bunny-sdk-python/issues"
Changelog = "https://github.com/jlarmstrongiv/bunny-sdk-python/commits/main/"

[tool.poetry.dependencies]
python = "^3.8"
microsoft-kiota-abstractions = "^1.3.3"
microsoft-kiota-http = "^1.3.3"
microsoft-kiota-serialization-json = "^1.3.0"
microsoft-kiota-serialization-text = "^1.0.0"
microsoft-kiota-serialization-form = "^0.1.0"
microsoft-kiota-serialization-multipart = "^0.1.0"
urllib3 = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
